Moved 3 links widgets from list_list_item.dart to list_details.dart
