+++ lib/lists/list_list_item.dart
@@ -6,9 +6,7 @@ import 'package:screensite/lists/lists_page.dart';
 import 'package:screensite/providers/firestore.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:jiffy/jiffy.dart';
-import 'package:url_launcher/url_launcher.dart';
 import 'package:flutter/gestures.dart';
-
 import '../extensions/string_validations.dart';
 import '../search/search_details.dart';
 
@@ -50,112 +48,6 @@ class ListItem extends ConsumerWidget {
                     ref.read(activeList.notifier).value = entityId;
                   },
                 ),
-                Row(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: <Widget>[
-                      Container(
-                          padding: EdgeInsets.all(8.0),
-                          child: Text.rich(TextSpan(
-                              style: TextStyle(
-                                  decoration: TextDecoration.underline),
-                              //make link underline
-                              text: "View source page",
-                              recognizer: TapGestureRecognizer()
-                                ..onTap = () async {
-                                  //on tap code here, you can navigate to other page or URL
-                                  final url = Uri.parse(
-                                      entityDoc.data()!['domainURL'] ?? '#');
-                                  var urllaunchable = await canLaunchUrl(
-                                      url); //canLaunch is from url_launcher package
-                                  if (urllaunchable) {
-                                    await launchUrl(
-                                        url); //launch is from url_launcher package to launch URL
-                                  } else {
-                                    print("URL can't be launched.");
-                                  }
-                                })))
-                    ]),
-                Row(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: <Widget>[
-                      Container(
-                          padding: EdgeInsets.all(8.0),
-                          child: Text.rich(TextSpan(
-                              style: TextStyle(
-                                  decoration: TextDecoration.underline),
-                              //make link underline
-                              text: "View source list",
-                              recognizer: TapGestureRecognizer()
-                                ..onTap = () async {
-                                  //on tap code here, you can navigate to other page or URL
-                                  final url = Uri.parse(
-                                      entityDoc.data()!['listURL'] ?? '#');
-                                  var urllaunchable = await canLaunchUrl(
-                                      url); //canLaunch is from url_launcher package
-                                  if (urllaunchable) {
-                                    await launchUrl(
-                                        url); //launch is from url_launcher package to launch URL
-                                  } else {
-                                    print("URL can't be launched.");
-                                  }
-                                })))
-                    ]),
-                InkWell(
-                    child: IconButton(
-                        icon: const Icon(Icons.edit),
-                        onPressed: () {
-                          showDialog(
-                              context: context,
-                              builder: (BuildContext context) {
-                                final _formKey = GlobalKey<FormState>();
-                                return AlertDialog(
-                                  scrollable: true,
-                                  title: Text('Sanction list entity fields'),
-                                  content: Padding(
-                                    padding: const EdgeInsets.all(8.0),
-                                    child: Form(
-                                      key: _formKey,
-                                      child: Column(
-                                        children: <Widget>[
-                                          DocFieldTextEdit(
-                                              FirebaseFirestore.instance
-                                                  .doc('list/${entityId}'),
-                                              'name',
-                                              decoration: InputDecoration(
-                                                  hintText: "Entity Name")),
-                                          DocFieldTextEdit(
-                                              FirebaseFirestore.instance
-                                                  .doc('list/${entityId}'),
-                                              'address',
-                                              decoration: InputDecoration(
-                                                  hintText: "Entity address")),
-                                          DocFieldTextEdit(
-                                              FirebaseFirestore.instance
-                                                  .doc('list/${entityId}'),
-                                              'dataSource',
-                                              decoration: InputDecoration(
-                                                  hintText: "Data Source")),
-                                          DocFieldTextEdit(
-                                              FirebaseFirestore.instance
-                                                  .doc('list/${entityId}'),
-                                              'website',
-                                              decoration: InputDecoration(
-                                                  hintText: "Website")),
-                                        ],
-                                      ),
-                                    ),
-                                  ),
-                                  actions: [
-                                    ElevatedButton(
-                                      onPressed: () {
-                                        Navigator.of(context).pop();
-                                      },
-                                      child: const Text('Done'),
-                                    )
-                                  ],
-                                );
-                              });
-                        })),
               ])));
   }
